+++ lib/commands/init.js
@@ -1,92 +1,69 @@
 var Q = require('q');
+var _ = require('underscore');
 var fs = require('fs');
-var readline = require('readline');
-var colors = require('colors');
-var OAuth2Client = require('googleapis').auth.OAuth2;
+var path = require('path');
+var request = require('request');
+var google = require('googleapis');
+var authenticate = require('../utils/authenticate');
+var clone = require('../utils/clone');
 var config = require('../config');
-var error = moduleErrors();
-var auth = {}; // Persisted client credentials
+var projectUtils = require('../utils/project');
 
-var GOOGLE_AUTH_SCOPE = ['https://www.googleapis.com/auth/drive','https://www.googleapis.com/auth/drive.scripts'];
+var DOWNLOAD_URL = 'https://script.google.com/feeds/download/export?format=json&id=';
 
-module.exports = function(path) {
-  getCredentialsFromFile(path)
-    .then(authenticateWithGoogle)
-    .then(saveAuthenticationConfig)
-    .then(shutdownSafely)
+module.exports = function(fileId) {
+  return projectUtils.getProjectById(fileId)
+    .then(removeCodeGs)
+    .then(function(project) {
+      project.path = process.cwd()
+      return project
+    })
+    .then(writeManifest)
+    .then(addProjectFiles)
+    .then(function(project) {
+      console.log('All done!');
+      return project;
+    })
     .catch(function(err) {
-      console.log('Error running init command', err);
+      console.log('Error running download command', err);
     });
 }
 
-function getCredentialsFromFile(path) {
+function removeCodeGs(project) {
   var deferred = Q.defer();
-  fs.readFile(path, function(err, res) {
-    if (err) return deferred.reject(error.wrongCredentialPath);
-    if (!res) return deferred.reject();
+  project.files = _.filter(project.files, function(file) {
+    return !(file.name === 'Code' && file.type === 'server_js')
+  });
+  deferred.resolve(project);
+  return deferred.promise;
+}
 
-    var credentials = JSON.parse(res);
-    if (!credentials.web) return deferred.reject(error.wrongCredentialsFile);
-
-    // Add important auth information to persist
-    auth.client_id = credentials.web.client_id;
-    auth.client_secret = credentials.web.client_secret;
-    auth.redirect_uri = credentials.web.redirect_uris[0];
-
-    return deferred.resolve(credentials.web);
+// Persist project in JSON format so we can update easily
+function writeManifest(project) {
+  var deferred = Q.defer();
+  var filepath = path.join(project.path, config.MANIFEST_NAME);
+  fs.writeFile(filepath, JSON.stringify(project, "", 2), function(err) {
+    if (err) return deferred.reject(err);
+    deferred.resolve(project);
   });
   return deferred.promise;
 }
 
-function authenticateWithGoogle(credentials) {
-  var deferred = Q.defer();
-  var rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  });
-
-  var oauth2Client = new OAuth2Client(credentials.client_id, credentials.client_secret, credentials.redirect_uris[0]);
-  var url = oauth2Client.generateAuthUrl({
-    access_type: 'offline', // Required to receive a refresh token
-    approval_prompt: 'force', // Forces consent re-prompt to guarentee refresh_token in response
-    scope: GOOGLE_AUTH_SCOPE
-  });
-
-  console.log('\nPlease visit the following url in your browser (you\'ll only have to do this once):'.cyan, url.green);
-  console.log('\nLook in the url for ?code=XXXXXXXXX and copy everything after the \'=\''.yellow)
-  rl.question('Enter the code here: ', getAccessToken);
-
-  function getAccessToken(code) {
-    oauth2Client.getToken(code, function(err, tokens) {
+function addProjectFiles(project) {
+  var writeFilePromises = project.files.map(function(file) {
+    var deferred = Q.defer();
+    var filepath = path.join(project.path, file.name + getFileExtension(file));
+    fs.writeFile(filepath, file.source, function(err) {
       if (err) return deferred.reject(err);
-      auth.refresh_token = tokens.refresh_token;
       deferred.resolve();
     });
-  };
-  return deferred.promise;
+    return deferred.promise;
+  });
+  return Q.all(writeFilePromises).then(_.constant(project));
 }
 
-function saveAuthenticationConfig() {
-  var deferred = Q.defer();
-  fs.writeFile(config.STORAGE_FILE, JSON.stringify(auth, "", 2), function(err, res) {
-    if (err) return deferred.reject(err);
-    deferred.resolve();
-  })
-  return deferred.promise;
+function getFileExtension(file) {
+  if (file.type === 'server_js') return '.js';
+  if (file.type === 'html') return '.html';
+  throw new Error('Unsupported file type found');
 }
-
-function shutdownSafely(){
-  console.log('Authentication successful! Ready to sync.'.green);
-  process.stdin.destroy();
-  process.exit(0);
-};
-
-function moduleErrors() {
-  return {
-    pathNotFound: 'Credentials path not found. Please input a path to your downloaded JSON credentials and try again.',
-    wrongCredentialPath: 'Could not read path to credentials file. Please check your path and try again',
-    credentialsNotFound: 'Credentials not found. Please check your path and try again',
-    wrongCredentialsFile: 'Path did not include correct credentials. Please check that you downloaded the right JSON credentials.'
-  }
-}
-
