+++ lib/commands/upload.js
@@ -6,7 +6,7 @@ var google = require('googleapis');
 var authenticate = require('../utils/authenticate');
 var updateManifest = require('../utils/manifestor');
 var config = require('../config');
-var error = moduleErrors();
+var errors = require('../utils/errors');
 
 module.exports = function(path) {
   var projectPath = _.isString(path) ? path : process.cwd();
@@ -29,7 +29,7 @@ function getManifestFromProject(projectPath) {
 
     var manifest = JSON.parse(res);
     manifest.path = projectPath;
-    if (!manifest.files) return deferred.reject(error.wrongManifest);
+    if (!manifest.files) return deferred.reject(errors.wrongManifest);
     deferred.resolve(manifest)
   });
   return deferred.promise;
@@ -53,13 +53,4 @@ function sendToGoogle(manifest) {
     })
   });
   return deferred.promise;
-}
-
-function moduleErrors() {
-  return {
-    pathNotFound: 'Project path not found. Double check your path and try again',
-    wrongManifest: 'The provided path doesn\'t seem to be an imported Google Apps Script project.\n' +
-                   'Make sure to you have input the proper path to your project and try again.\n' +
-                   'If you haven\'t imported a project yet, run import.js first'
-  };
 }
