+++ lib/utils/authenticate.js
