+++ lib/utils/clone.js
