+++ lib/utils/errors.js
@@ -1,0 +1,12 @@
+module.exports = {
+  fileIdNotFound: 'Script file ID not found. Please input an ID and try again.',
+  noFiles: 'Looks like there are no files associated with this project. Check the id and try again.',
+  pathNotFound: 'Credentials path not found. Please input a path to your downloaded JSON credentials and try again.',
+  wrongCredentialPath: 'Could not read path to credentials file. Please check your path and try again',
+  credentialsNotFound: 'Credentials not found. Please check your path and try again',
+  wrongCredentialsFile: 'Path did not nclude correct credentials. Please check that you downloaded the right JSON credentials.',
+  pathNotFound: 'Project path not found. Double check your path and try again',
+  wrongManifest: 'The provided path doesn\'t seem to be an imported Google Apps Script project.\n' +
+                 'Make sure to you have input the proper path to your project and try again.\n' +
+                 'If you haven\'t imported a project yet, run import.js first',
+};
