+++ lib/utils/manifestor.js
@@ -4,7 +4,7 @@ var dir = require('node-dir');
 var path = require('path');
 var config = require('../config');
 
-// Readies manifest for export by updating with all local file changes 
+// Readies manifest for export by updating with all local file changes
 module.exports = function(manifest) {
   var deferred = Q.defer();
